Merge pull request #1245 from digitalinfinity:lm_safe

Payload:

Several enhancements to build.sh and runtests.py
Updates to CMakeLists
Update to jenkins CI to run unit tests as part of CI builds
Support for ch to be statically linked with ChakraCore libraries on Linux/OS X
Updates to JSRT APIs to support Utf8 (ch uses only the xplat API surface of Jsrt)
Props files for building ChakraCore with VC++ clang frontend on Windows
Cleanup of __declspec(noinline)
Protected some more code under #ifdef ENABLE_SIMDJS
Platform Agnostic implementation of DateTime related functions
Add support for "Baseline" mode while running unit tests to reduce noise in tests that have platform-specific offset outputs
Several bug fixes- unit tests are now clean!

+++ lib/Backend/GlobOptExpr.cpp
@@ -5,8 +5,8 @@
 #include "Backend.h"
 
 
-uint8 OpCodeToHash[Js::OpCode::Count];
-static Js::OpCode HashToOpCode[Js::OpCode::Count];
+uint8 OpCodeToHash[(uint)Js::OpCode::Count];
+static Js::OpCode HashToOpCode[(uint)Js::OpCode::Count];
 
 class CSEInit
 {
